+++ src/components/script/ScriptEditorEvent.tsx
@@ -154,6 +154,17 @@ const ScriptEditorEvent = React.memo(
       [dispatch, id, parentId, parentKey, parentType, scriptEventSelectionIds],
     );
 
+    const onToggleSelection = useCallback(() => {
+      dispatch(
+        editorActions.toggleScriptEventSelectedId({
+          scriptEventId: id,
+          parentId,
+          parentKey,
+          parentType,
+        }),
+      );
+    }, [dispatch, id, parentId, parentKey, parentType]);
+
     const onFetchClipboard = useCallback(
       (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement>) => {
         onSelect(e.shiftKey);
@@ -337,6 +348,62 @@ const ScriptEditorEvent = React.memo(
       }
     }, [context.entityType, context.instanceId, dispatch, id]);
 
+    const contextMenuKeyboardHandler = useCallback(
+      (e: React.KeyboardEvent<HTMLElement> | KeyboardEvent) => {
+        if ((e.ctrlKey || e.metaKey) && e.code === "KeyG") {
+          e.stopPropagation();
+          e.preventDefault();
+          dispatch(
+            entitiesActions.groupScriptEvents({
+              scriptEventIds: scriptEventSelectionIds,
+              parentId,
+              parentKey,
+              parentType,
+            }),
+          );
+          return true;
+        } else if ((e.ctrlKey || e.metaKey) && e.code === "KeyC") {
+          e.stopPropagation();
+          e.preventDefault();
+          dispatch(
+            clipboardActions.copyScriptEvents({
+              scriptEventIds:
+                scriptEventSelectionIds.length > 0
+                  ? scriptEventSelectionIds
+                  : [id],
+            }),
+          );
+          return true;
+        } else if ((e.ctrlKey || e.metaKey) && e.code === "KeyV") {
+          e.stopPropagation();
+          e.preventDefault();
+          dispatch(
+            clipboardActions.pasteScriptEvents({
+              entityId: parentId,
+              type: parentType,
+              key: parentKey,
+              insertId: id,
+              before: e.shiftKey,
+            }),
+          );
+          return true;
+        } else if ((e.ctrlKey || e.metaKey) && e.code === "Slash") {
+          e.stopPropagation();
+          e.preventDefault();
+          dispatch(
+            entitiesActions.toggleScriptEventComment({
+              scriptEventId: id,
+              additionalScriptEventIds: scriptEventSelectionIds,
+            }),
+          );
+          return true;
+        }
+
+        return false;
+      },
+      [dispatch, id, parentId, parentKey, parentType, scriptEventSelectionIds],
+    );
+
     const contextMenuItems = useMemo(
       () =>
         scriptEvent
@@ -442,24 +509,13 @@ const ScriptEditorEvent = React.memo(
 
     const onKeyDown = useCallback(
       (e: KeyboardEvent) => {
-        // Group selection with ctrl/cmd + g
-        if (
-          (e.ctrlKey || e.metaKey) &&
-          e.key === "g" &&
-          scriptEventSelectionIds[0] === id
-        ) {
-          dispatch(
-            entitiesActions.groupScriptEvents({
-              scriptEventIds: scriptEventSelectionIds,
-              parentId,
-              parentKey,
-              parentType,
-            }),
-          );
-          return;
+        if (contextMenu) {
+          if (contextMenuKeyboardHandler(e)) {
+            onContextMenuClose();
+          }
         }
       },
-      [dispatch, id, parentId, parentKey, parentType, scriptEventSelectionIds],
+      [contextMenu, contextMenuKeyboardHandler, onContextMenuClose],
     );
 
     useEffect(() => {
@@ -467,7 +523,7 @@ const ScriptEditorEvent = React.memo(
       return () => {
         window.removeEventListener("keydown", onKeyDown);
       };
-    });
+    }, [onKeyDown]);
 
     if (!scriptEvent) {
       return null;
@@ -518,9 +574,11 @@ const ScriptEditorEvent = React.memo(
               isBreakpoint={breakpointEnabled}
               breakpointTitle={l10n("FIELD_BREAKPOINT")}
               menuItems={contextMenuItems}
+              menuKeyboardHandler={contextMenuKeyboardHandler}
               onOpenMenu={onFetchClipboard}
               onContextMenu={onContextMenu}
               onToggle={!rename ? toggleOpen : undefined}
+              onToggleSelection={onToggleSelection}
             >
               {isVisible && (
                 <>